requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==2.0.0
python-decouple==3.3
django-rest-swagger==2.2.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.8
pyparsing==2.4.7
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
